==> DSM-BCD/NOTEBOOK/ALGORITMOS/1_EDA.ipynb <==
# Data Science Methodology for Breast Cancer Diagnosis (DSM-BCD)

**Problema.** Colombia presenta limitaciones con respecto al acceso de la detección y el diagnóstico temprano del cáncer, provocado en la mayoría de los casos por factores como el estrato socio-económico, la cobertura del seguro de salud, el origen y la accesibilidad. En promedio, el tiempo de espera de un paciente es de 90 días desde la aparición de los síntomas hasta el diagnóstico de dicho cáncer. La primera acción para reducir la tasa de mortalidad por cáncer de mama debe estar enfocada en la agilidad del diagnóstico y el acceso oportuno a la atención.  Así, el objetivo de esta investigación es aplicar las etapas de la metodología KDD\footnote{Knowledge Discovery in Databases} al conjunto de datos de morbilidad por cáncer entre los años 2019 y 2020 en el municipio de Pereira-Risaralda. Esto con la finalidad de pronosticar y caracterizar el tipo de población mas susceptible de padecer esta enfermedad según su edad, genero, zona y régimen de salud.

**Conjunto de datos.** Un total de 817 muestras de tumores de mama se perfilaron con cinco plataformas diferentes como se ha descrito previamente (Cancer Genome Atlas Research Network, 2014) y también se perfilaron 633 casos mediante matriz de proteínas en fase inversa (RPPA)


# Análisis exploratorio de datos

# Importar librerias
import pandas as pd
import numpy as np
from IPython.display import Image
import matplotlib.pyplot as plt
from dataprep.eda import plot, plot_correlation, plot_missing,create_report ,configs
from dataprep.clean import clean_headers
from dataprep.clean import clean_duplication
from dataprep.clean import clean_date
from dataprep.clean import clean_text

# Abrir conjunto de datos delimitados por espacios
with open('brca_tcga_pub2015_clinical_data.csv') as f:
    breast_cancer=pd.read_csv(f, delimiter="\t")

# Generar una copia del data-set original
bc = breast_cancer.copy()
bc.shape

#missing=plot_missing(bc,config={'height': 600, 'width': 600})
#missing.show_browser()

## 1. Análisis parcial de datos crudos

#report = create_report(breast_cancer, title='Breast Invasive Carcinoma (TCGA, Cell 2015)' , config={'bar.sort_descending': True, 'bar.yscale': 'linear', 'height': 100, 'width': 400, 'bar.color': '#FF8C00', 'hist.color': '#FF8C00'})

#report.show_browser()

#report.save()

#f=plot(bc,config={'bar.sort_descending': True, 'bar.yscale': 'linear', 'height': 100, 'width': 400, 'bar.color': '#FF8C00','bar.enable':True, 'hist.color': '#FF8C00'})
#f.show_browser()


## 2. Transformación de  variables

### 2.1 Re-nombrar Columnas

# Estandarizar cabecera en format snake
bc=clean_headers(breast_cancer)

# Reemplazar nomenclatura extensa
bc.columns=bc.columns.str.replace('american_joint_committee_on_cancer_','')

bc.columns=bc.columns.str.replace('international_classification_of_diseases_for_oncology_third_edition_icd_o_3_','')

# Renombrar variables con un nombre corto
bc.rename(
columns={
'neoplasm_disease_lymph_node_stage_code':'neoplasm_lymph_code',
'neoplasm_lymph_node_stage_code':'neoplasm_stage',
'brachytherapy_first_reference_point_administered_total_dose':'brachytherapy',
'birth_from_initial_pathologic_diagnosis_date':'birth_initial_diagnosis',
'death_from_initial_pathologic_diagnosis_date':'death_initial_diagnosis',
'last_alive_less_initial_pathologic_diagnosis_date_calculated_day_value':'last_alive_date',
'neoadjuvant_therapy_type_administered_prior_to_resection_text':'neoadjuvant_therapy',
'prior_cancer_diagnosis_occurence':'prior_diagnosis_occurence',
'informed_consent_verified':'consent_verified',
'primary_lymph_node_presentation_assessment_ind_3':'lymph_presentation',
'positive_finding_lymph_node_hematoxylin_and_eosin_staining_microscopy_count':'positive_lymph_hematoxylin',
'positive_finding_lymph_node_keratin_immunohistochemistry_staining_method_count':'positive_lymph_keratin',
'lymph_node_s_examined_number':'lymph_examined_number',
'first_pathologic_diagnosis_biospecimen_acquisition_method_type':'biospecimen_method',
'first_pathologic_diagnosis_biospecimen_acquisition_other_method_type':'biospecimen_other_method',
'new_neoplasm_event_post_initial_therapy_indicator':'new_neoplasm_event',
'adjuvant_postoperative_pharmaceutical_therapy_administered_indicator':'pharmaceutical_therapy',
'tissue_prospective_collection_indicator':'tissue_prospective_indicator',
'did_patient_start_adjuvant_postoperative_radiotherapy':'postoperative_radiotherapy',
'tissue_retrospective_collection_indicator':'tissue_retrospective_indicator',
'number_of_samples_per_patient':'number_samples',
'surgery_for_positive_margins':'surgery_positive',
'surgery_for_positive_margins_other':'surgery_positive_other',
'surgery_for_positive_margins_other':'surgery_positive_other',
'neoplasm_histologic_type_name':'neoplasm_histologic_type',
'tumor_other_histologic_subtype':'tumor_other_subtype',
'year_cancer_initial_diagnosis':'year_initial_diagnosis',
'first_surgical_procedure_other':'surgical_other'
}, inplace=True)

# Reemplazar artículos, pronombres, preposiciones
bc.columns=bc.columns.str.replace('_to_','_')

bc.columns=bc.columns.str.replace('_and_','_')

bc.columns=bc.columns.str.replace('_by_','_')

### 2.2 Estandarizar variables de Texto

# Procesar datos para que queden en minuscula
custom_pipeline = [{"operator": "lowercase"}]
for i in bc.columns:
    bc=clean_text(bc,i,pipeline=custom_pipeline)

# Estandarizar variables tipo NaN
bc=bc.replace(
    ["<NA>",
     "nan",
     "nan"
     ],[np.NaN,
    np.NaN,
    np.NaN
    ],regex=True)

# Estandarizar variable brachytherapy
bc.brachytherapy=bc.brachytherapy.replace(
    ['no value given',
    '% ihc',
    '-',
    ' '],
    [np.NaN,
    'ihc',
    '',
    ''],regex=True)

# Estandarizar variable publication_version_type
bc.publication_version_type=bc.publication_version_type.replace(
    ['th|rd'],[''],regex=True)

# Estandarizar variable cent_17_copy_number
bc.cent_17_copy_number=bc.cent_17_copy_number.replace(
    ['polisomy',],[np.NaN,],regex=True)

# Estandarizar variable er_positivity_scale_used
bc.er_positivity_scale_used=bc.er_positivity_scale_used.replace(
    ['point scale',],
    [''],regex=True)

# Estandarizar variable disease_free_status
bc.disease_free_status=bc.disease_free_status.replace(
    ['0:diseasefree',
    '1:recurred/progressed'],
    ['diseasefree',
    'progressed'],regex=True)

# Estandarizar variable days_last_followup
bc.days_last_followup = bc.days_last_followup.replace(
    ['-'],[''],regex=True)

# Estandarizar variable her_2_copy_number
bc.her_2_copy_number=bc.her_2_copy_number.replace(
    ['<',
    '>',
    'not amplified'
    ],
    ['',
    '',
    0],regex=True)

# Estandarizar variable er_positivity_scale_other
bc.er_positivity_scale_other=bc.er_positivity_scale_other.replace(
    ['protein',
    'allred score 0',
    '=',
    'scrore',
    'h-score',
    'intensity',
    'strong using weak, moderate and strong',
    'moderate using the scale of weak, moderate, strong',
    'moderate using scale of weak, moderate, strong',
    ' \(per outside facility\)',
    '  ',
    ' '
    ],
    ['',
    'allred score',
    '',
    'score',
    'hscore',
    '',
    'strong',
    'moderate',
    'moderate',
    '',
    '',
    ''
    ],regex=True)

# Estandarizar variable er_status_ihc
bc.er_status_ihc=bc.er_status_ihc.replace(
    ['indeterminate'],
    ['positive'],regex=True)

# Estandarizar variable surgical_other
bc.surgical_other=bc.surgical_other.replace(
    ['total mastectomy',
     'modified radical masectomy',
    'biospy',
    "patey's",
    'lymoh'],
    ["patey's surgery",
     "patey's surgery",
    'biopsy',
    'pateys',
    'lymph'],regex=True)

bc.surgical_other=bc.surgical_other.replace(
    ['biospy',
    "patey's",
    'lymoh'],
    ['biopsy',
    'pateys',
    'lymph'],regex=True)

# Estandarizar variable her_2_fish_status
bc.her_2_fish_status=bc.her_2_fish_status.replace(
    ['equivocal',
     'indeterminate'],
    ['negative',
    'negative'],regex=True)

# Estandarizar variable neoplasm_histologic_type
bc.neoplasm_histologic_type=bc.neoplasm_histologic_type.replace(
    ["(please specify)",
    "other, specify",
    '\(','\)'],
    ['',
    'other',
    '',''],regex=True)

# Estandarizar variable tumor_other_subtype
bc.tumor_other_subtype=bc.tumor_other_subtype.replace(
    ['ductal and lobular features',
    'infiltrating ductal & lobular',
    'lobular carcinoma with ductal features',
    'lobular/ductal'],
    ['ductal/lobular',
    'infiltrating ductal and lobular carcinoma',
    'ductal/lobular',
    'ductal/lobular'],regex=True)

# Estandarizar variable menopause_status
bc.menopause_status=bc.menopause_status.replace(
    ['post \(prior bilateral ovariectomy or >12 mo since lmp with no prior hysterectomy\)',
    'pre \(<6 months since lmp and no prior bilateral ovariectomy and not on estrogen replacement\)',
    'indeterminate \(neither pre or postmenopausal\)',
    'peri \(6-12 months since last menstrual period\)'
    ],
    ['post',
    'pre',
    'peri',
    'peri'],regex=True)

# Estandarizar variable metastatic_site
bc.metastatic_site=bc.metastatic_site.replace(
    ['lung\|bone\|liver\|other, specify',
     'other, specify',
    'bone\|liver',
    ],
    ['other',
    'other',
    'bone-liver',],regex=True)

# Estandarizar variable biospecimen_method
bc.biospecimen_method=bc.biospecimen_method.replace(
    ['other method, specify:',
    'cytology \(e.g. peritoneal or pleural fluid\)'
    ],
    ['other',
    'cytology'
    ],regex=True)

# Estandarizar variable biospecimen_other_method
bc.biospecimen_other_method=bc.biospecimen_other_method.replace(
    ['biopsy, nos',
    "patey's",
    'surgery'
    ],
    ['biopsy',
    "pateys",
    'suregery'
    ],regex=True)

# Estandarizar variable overall_survival_status
bc.overall_survival_status=bc.overall_survival_status.replace(
    ['0:',
    '1:'],
    ['',
    ''],regex=True)

# Estandarizar variable tissue_prospective_indicator
bc.tissue_prospective_indicator=bc.tissue_prospective_indicator.replace(
    ['yes',
     'no'],
    ['prospective',
    'retrospective'],regex=True)

# Estandarizar variable pr_positivity_define_method
bc.pr_positivity_define_method=bc.pr_positivity_define_method.replace(
    ['no value given',
    '%ihc',
    'per outside facility report',
    '-',
    ' ',
    ],[np.NaN,
    'ihc',
    '',
    '',
    ''],regex=True)

# Estandarizar variable pr_positivity_scale_other
bc.pr_positivity_scale_other=bc.pr_positivity_scale_other.replace(
    ['protein',
    'allred score 0',
    '=',
    'scrore',
    'h-score',
    'intensity',
    'strong using weak, moderate and strong',
    'moderate using the scale of weak, moderate, strong',
    'moderate using scale of weak, moderate, strong',
    ' \(per outside facility\)',
    'per outside facility report',
    'strong, using scale of weak, moderate and strong',
    'allread',
    '  ',
    ' ',
    ],
    ['',
    'allred score',
    '',
    'score',
    'hscore',
    '',
    'strong',
    'moderate',
    'moderate',
    '',
    '',
    'Strong',
    'allred',
    '',
    '',
    ],regex=True)

# Estandarizar variable staging_system
bc.staging_system=bc.staging_system.replace(
    ['other \(specify\)'
    ],
    ['other'
    ],regex=True)

# Estandarizar variable staging_system_1
bc.staging_system_1=bc.staging_system_1.replace(
    ['sln and non-sln bx|'+
     'sln and non-sln biopsy|'+
     'sentinel ln and one non sentinel ln|'
     'sentinel \+ non sentinel|'
     'sn\+1 non sentinel node|'
     'sentinel lymph node biopsy and non-sentinel lymph node biopsy'
    ],
    ['sentinel lymph node and non-sentinel lymph node biopsy'
    ],regex=True)

# Estandarizar variable pr_positivity_ihc_intensity_score
bc.pr_positivity_ihc_intensity_score=bc.pr_positivity_ihc_intensity_score.replace(
    ['\+'],[''],regex=True)

# Estandarizar variable pr_positivity_scale_used
bc.pr_positivity_scale_used=bc.pr_positivity_scale_used.replace(
    ['point scale',],
    [''],regex=True)

# Estandarizar variable pr_status_ihc
bc.pr_status_ihc=bc.pr_status_ihc.replace(
    ['indeterminate'],
    ['positive'],regex=True)

# Estandarizar variable primary_tumor_site
bc.primary_tumor_site=bc.primary_tumor_site.replace(
    ['left',
    'left upper outer quadrant',
    'left upper inner quadrant',
    'right',
    'right upper outer quadrant',
    'right upper outer quadrant|right',
    'right lower outer quadrant|right',
    'right lower outer quadrant',
    'left lower outer quadrant',
    'right upper inner quadrant',
    'right lower inner quadrant',
    'left upper inner quadrant|left upper outer quadrant|left lower inner quadrant|left lower outer quadrant',
    'left upper inner quadrant|left upper outer quadrant',
    'right|right upper outer quadrant',
    'left upper outer quadrant|left',
    'left lower inner quadrant',
    'right upper inner quadrant|right lower inner quadrant',
    'left upper outer quadrant|left lower outer quadrant',
    'left upper outer quadrant|left lower outer quadrant|left',
    'left lower inner quadrant|left lower outer quadrant|left',
    'right upper inner quadrant|right upper outer quadrant',
    'right lower inner quadrant|right lower outer quadrant',
    'right upper outer quadrant|right lower outer quadrant',
    'left lower outer quadrant|left',
    'right upper inner quadrant|right',
    'left lower inner quadrant|left',
    'left upper inner quadrant|left upper outer quadrant|left',
    'right upper inner quadrant|right upper outer quadrant|right',
    'right lower inner quadrant|right lower outer quadrant|right',
    'left upper inner quadrant|left upper outer quadrant|left lower inner quadrant|left lower outer quadrant|left',
    'right upper inner quadrant|right lower inner quadrant|right',
    'left upper inner quadrant|left lower inner quadrant|left lower outer quadrant|left',
    'right upper inner quadrant|right upper outer quadrant|right lower inner quadrant|right lower outer quadrant|right',
    'left upper inner quadrant|left',
    'right upper outer quadrant|right lower inner quadrant|right lower outer quadrant',
    'right|right upper inner quadrant',
    'left upper inner quadrant|left lower inner quadrant',
    'left|left upper outer quadrant',
    'left lower inner quadrant|left lower outer quadrant',
    'left|left upper inner quadrant|left upper outer quadrant|left lower inner quadrant|left lower outer quadrant',
    'left|left upper inner quadrant',
    'right|right lower inner quadrant'
    ],
    ['C50.912',
    'C50.412',
    'C50.212',
    'C50.911',
    'C50.411',
    'C50.411',
    'C50.511',
    'C50.511',
    'C50.512',
    'C50.211',
    'C50.311',
    'C50.912',
    'C50.912',
    'C50.411',
    'C50.412',
    'C50.312',
    'C50.911',
    'C50.912',
    'C50.912',
    'C50.912',
    'C50.911',
    'C50.911',
    'C50.911',
    'C50.512',
    'C50.211',
    'C50.312',
    'C50.912',
    'C50.911',
    'C50.911',
    'C50.912',
    'C50.911',
    'C50.912',
    'C50.911',
    'C50.212',
    'C50.911',
    'C50.211',
    'C50.912',
    'C50.412',
    'C50.912',
    'C50.912',
    'C50.212',
    'C50.311'
    ])

### 2.3 Ajustar tipo de variables

# Crear vector con las variables a transformar
convert_data = ['publication_version_type','cent_17_copy_number','er_positivity_scale_used',
'her_2_copy_number','her_2_ihc_score','ihc_score','positive_lymph_keratin',
'pr_positivity_ihc_intensity_score','pr_positivity_scale_used','fraction_genome_altered',
'number_samples','tmb_nonsynonymous','her_2_cent_17_ratio','lymph_examined_number','mutation_count',
'overall_survival_months','disease_free_months','days_last_followup']

# Convertir variables numericas a su tipo correspondiente.
for i in convert_data:
    bc[i] = pd.to_numeric(bc[i].fillna(0))
    print(i,':',bc[i].dtype)

## 3. Tratamiento de datos ausentes

### 3.1 Detección de datos ausentes

##missing=plot_missing(bc,config={'height': 600, 'width': 600})
##missing.show_browser()

### 3.2 Imputación de datos ausentes


from sklearn.impute import KNNImputer

# Imputar variables numericas
imputer = KNNImputer(n_neighbors=5, weights="distance")
#imputer = SimpleImputer(strategy='mean', missing_values=np.nan)

imput_data=['diagnosis_age',
            'cent_17_copy_number',
            'birth_initial_diagnosis',
            'days_sample_collection',
            'her_2_cent_17_ratio',
            'disease_free_months',
            'days_last_followup',
            'year_initial_diagnosis',
            'positive_lymph_hematoxylin',
            'lymph_examined_number',
            'mutation_count',
            'overall_survival_months',
            'tmb_nonsynonymous'
]

for i in imput_data :
    bc[i]=bc[i].replace([0.0,0,'0','0.0'],[np.nan,np.nan,np.nan,np.nan])

    # Ajustamos el modelo e imputamos los missing values
    imputer.fit(bc[[i]])
    bc[i] = imputer.transform(bc[[i]]).ravel()
    bc[i] = round(bc[i],1)


from sklearn.impute import SimpleImputer

# Imputar variables categoricas
mode_data=['neoplasm_disease_stage_code',
           'brachytherapy',
           'publication_version_type',
           'er_positivity_scale_used',
           'cancer_type_detailed',
           'disease_free_status',
           'er_positivity_scale_other',
           'er_status_ihc_percent_positive',
           'ethnicity_category',
           'surgical_other',
           'er_status_ihc',
           'her_2_fish_status',
           'her_2_ihc_score',
           'her_2_ihc_percent_positive',
           'neoplasm_histologic_type',
           'neoadjuvant_therapy',
           'prior_diagnosis_occurence','ihc_her_2','ihc_score',
           'lymph_presentation',
           'menopause_status',
           'metastatic_tumor_indicator',
           'biospecimen_method',
           'micromet_detection_ihc',
           'oct_embedded',
           'disease_surgical_margin_status',
           'primary_tumor_site',
           'tissue_prospective_indicator',
           'pr_positivity_define_method',
           'pr_positivity_scale_other',
           'positive_lymph_keratin',
           'pr_status_ihc',
           'pr_status_ihc_percent_positive',
           'pr_positivity_ihc_intensity_score',
           'pr_positivity_scale_used',
           'race_category','tissue_retrospective_indicator',
           'staging_system',
           'surgical_procedure_first',
           'tissue_source_site','person_neoplasm_status'
           ]

for i in mode_data :
    bc[i]=bc[i].replace([0.0,0,'0','0.0'],[np.nan,np.nan,np.nan,np.nan])
    bc[i] = bc[i].fillna(bc[i].mode()[0])
    imputer = SimpleImputer(strategy='most_frequent', 
                            missing_values=np.nan)
    imputer = imputer.fit(bc[[i]])
    bc[[i]] = imputer.transform(bc[[i]])

### 3.3 Eliminar datos innecesarios

#Eliminar variables inncesarias
bc=bc.drop(
[
'study_id',
'patient_id',
'sample_id',
'cancer_type',
'death_initial_diagnosis',
'last_alive_date',
'publication_version_type',
'disease_code',
'brachytherapy',
'form_completion_date',
'histology_code',
'site_code',
'cent_17_copy_number',
'birth_initial_diagnosis',
'pr_positivity_define_method',
'er_positivity_scale_used',
'er_positivity_scale_other',
'her_2_cent_17_cells_count',
'her_2_cent_17_scale_other',
'her_2_copy_number',
'her_2_cent_17_ratio',
'her_2_fish_method',
'her_2_positivity_method_text',
'her_2_positivity_scale_other',
'tumor_other_subtype',
'consent_verified',
'is_ffpe',
'ihc_score',
'margin_status_reexcision',
'prior_diagnosis_occurence',
'metastatic_site',
'metastatic_site_other',
'biospecimen_other_method',
'new_neoplasm_event',
'icd_10_classification',
'nte_cent_17_her_2_ratio',
'nte_er_ihc_intensity_score',
'nte_er_status',
'nte_er_status_ihc_positive',
'nte_her_2_fish_status',
'nte_her_2_positivity_ihc_score',
'nte_her_2_status',
'nte_her_2_status_ihc_positive',
'nte_pr_ihc_intensity_score',
'nte_pr_status_ihc',
'nte_pr_status_ihc_positive',
'other_patient_id',
'other_sample_id',
'pathology_report_file_name',
'pharmaceutical_therapy',
'project_code',
'positive_lymph_keratin',
'tissue_retrospective_indicator',
'pr_positivity_ihc_intensity_score',
'pr_status_ihc_percent_positive',
'pr_positivity_scale_other',
'pr_positivity_scale_used',
'postoperative_radiotherapy',
'number_samples',
'sample_type',
'somatic_status',
'staging_system_1',
'surgery_positive',
'surgery_positive_other',
'surgical_other',
'tissue_source_site',
'tumor_disease_anatomic_site',
'days_sample_collection',
'days_last_followup'
], axis=1
)

##missing=plot_missing(bc,config={'height': 400, 'width': 400})
##missing.show_browser()

## 4. Análisis Descriptivo 

report = create_report(bc, title='Breast Invasive Clean Report - 58 variables')
report.show_browser()

## 5. Análisis de correlación de variables 

#correlation = plot_correlation(bc)
#correlation.show_browser()

#Guardar CSV para modelos ML
bc.to_csv('Breast_Clean_Data.csv',index=False)
#bc.shape
