==> DSM-BCD/NOTEBOOK/ALGORITMOS/2_ALGORITMO_ML.ipynb <==
# Data Science Methodology for Breast Cancer Diagnosis (DSM-BCD)

**Problema.** Colombia presenta limitaciones con respecto al acceso de la detección y el diagnóstico temprano del cáncer, provocado en la mayoría de los casos por factores como el estrato socio-económico, la cobertura del seguro de salud, el origen y la accesibilidad. En promedio, el tiempo de espera de un paciente es de 90 días desde la aparición de los síntomas hasta el diagnóstico de dicho cáncer. La primera acción para reducir la tasa de mortalidad por cáncer de mama debe estar enfocada en la agilidad del diagnóstico y el acceso oportuno a la atención.  Así, el objetivo de esta investigación es aplicar las etapas de la metodología KDD\footnote{Knowledge Discovery in Databases} al conjunto de datos de morbilidad por cáncer entre los años 2019 y 2020 en el municipio de Pereira-Risaralda. Esto con la finalidad de pronosticar y caracterizar el tipo de población mas susceptible de padecer esta enfermedad según su edad, genero, zona y régimen de salud.

**Conjunto de datos.** Un total de 817 muestras de tumores de mama se perfilaron con cinco plataformas diferentes como se ha descrito previamente (Cancer Genome Atlas Research Network, 2014) y también se perfilaron 633 casos mediante matriz de proteínas en fase inversa (RPPA)



## Aprendizaje No supervisado

En el aprendizaje No supervisado, se trabaja con datos sin etiquetar cuya estructura es desconocida. El objetivo será la extracción de información significativa, sin la referencia de variables de salida conocidas, y mediante la exploración de la estructura de dichos datos sin etiquetar. En este caso de estudio se utilizará la siguiente técnica:

### Agrupamiento (Clustering)
El agrupamiento es una técnica exploratoria de análisis de datos, que se usa para organizar información en grupos sin tener conocimiento previo de su estructura. Cada grupo es un conjunto de objetos similares que se diferencia de los objetos de otros grupos. El objetivo es obtener un número de grupos de características similares. Los modelos de Clustering más conocidos son los siguientes:

#### K-means: 
Es un algoritmo diseñado para dividir datos no etiquetados en un cierto número (k) de agrupaciones distintas. En otras palabras, k-means encuentra observaciones que comparten características importantes, las une y las clasifica en grupos.

#### K-modes: 
El algoritmo de agrupación de k-modes es una extensión del
algoritmo k-means. Este algoritmo fue diseñado para agrupar grandes conjuntos de datos categóricos, y tiene como objetivo obtener las k modas que
representan a un conjunto de datos determinado.


## Desarrollo Clustering

import pandas as pd
import numpy as np
import seaborn as sns

with open('Breast_Clean_Data.csv') as f:
    dataset=pd.read_csv(f, delimiter=',')

dataset.head(10)

dataset.shape

data = dataset.sample(frac=0.95, random_state=42)
data_unseen = dataset.drop(data.index)

data.reset_index(drop=True, inplace=True)
data_unseen.reset_index(drop=True, inplace=True)

print('Data for Modeling: ' + str(data.shape))
print('Unseen Data For Predictions: ' + str(data_unseen.shape))

from pycaret.clustering import *

exp_clu101 = setup(data, normalize = True, 
                   session_id = 123)

models()

### K-Means Clustering

kmeans = create_model('kmeans')
kmeans_results = assign_model(kmeans)


plot_model(kmeans, plot='cluster')

plot_model(kmeans, plot = 'elbow')

plot_model(kmeans, plot = 'silhouette')

plot_model(kmeans, plot = 'tsne')

plot_model(kmeans, plot = 'distance')

plot_model(kmeans, plot = 'distribution')

kmeans_results = assign_model(kmeans)
kmeans_results
kmeans_results.groupby("Cluster").describe()

### Affinity Propagation	

ap = create_model('ap')

plot_model(ap,plot='cluster')

plot_model(ap, plot = 'tsne')

plot_model(ap, plot = 'distance')

plot_model(ap, plot = 'distribution')

ap_results = assign_model(ap)
ap_results
ap_results.groupby("Cluster").describe()

### Mean Shift Clustering	

meanshift = create_model('meanshift')

plot_model(meanshift,plot='cluster')

plot_model(meanshift, plot = 'tsne')

plot_model(meanshift, plot = 'distance')

plot_model(meanshift, plot = 'distribution')

meanshift_results = assign_model(meanshift)
meanshift_results
meanshift_results.groupby("Cluster").describe()

### Spectral Clustering

sc = create_model('sc')

plot_model(sc,plot='cluster')

plot_model(sc, plot = 'tsne')

plot_model(sc, plot = 'distribution')

sc_results = assign_model(sc)
sc_results
sc_results.groupby("Cluster").describe()

### Agglomerative Clustering

hclust = create_model('hclust')

plot_model(hclust,plot='cluster')

plot_model(hclust, plot = 'elbow')

plot_model(hclust, plot = 'tsne')

plot_model(hclust, plot = 'distribution')

hclust_results = assign_model(hclust)
hclust_results
hclust_results.groupby("Cluster").describe()



### Density-Based Spatial Clustering	

dbscan = create_model('dbscan')

plot_model(dbscan,plot='cluster')

plot_model(dbscan, plot = 'tsne')

plot_model(dbscan, plot = 'distribution')

dbscan_results = assign_model(dbscan)
dbscan_results
dbscan_results.groupby("Cluster").describe()

### OPTICS Clustering	

optics = create_model('optics')


plot_model(optics,plot='cluster')

plot_model(optics, plot = 'tsne')

plot_model(optics, plot = 'distribution')


optics_results = assign_model(optics)
optics_results
optics_results.groupby("Cluster").describe()


### Birch Clustering	

birch = create_model('birch')

plot_model(birch,plot='cluster')

plot_model(birch, plot = 'elbow')

plot_model(birch, plot = 'silhouette')

plot_model(birch, plot = 'tsne')


plot_model(birch, plot = 'distribution')

birch_results = assign_model(birch)
birch_results
birch_results.groupby("Cluster").describe()

### K-Modes Clustering	

kmodes = create_model('kmodes', num_clusters = 3)

plot_model(kmodes,plot='cluster')

plot_model(kmodes, plot = 'elbow')

plot_model(kmodes, plot = 'silhouette')

plot_model(kmodes, plot = 'tsne')

plot_model(kmodes, plot = 'distribution')

kmodes_results = assign_model(kmodes)
kmodes_results
kmodes_results.groupby("Cluster").describe()
