==> DSM-BCD/NOTEBOOK/ALGORITMOS/4_EDA_MODELO_CLUSTERING.ipynb <==
# Data Science Methodology for Breast Cancer Diagnosis (DSM-BCD)

**Problema.** Colombia presenta limitaciones con respecto al acceso de la detección y el diagnóstico temprano del cáncer, provocado en la mayoría de los casos por factores como el estrato socio-económico, la cobertura del seguro de salud, el origen y la accesibilidad. En promedio, el tiempo de espera de un paciente es de 90 días desde la aparición de los síntomas hasta el diagnóstico de dicho cáncer. La primera acción para reducir la tasa de mortalidad por cáncer de mama debe estar enfocada en la agilidad del diagnóstico y el acceso oportuno a la atención.  Así, el objetivo de esta investigación es aplicar las etapas de la metodología KDD\footnote{Knowledge Discovery in Databases} al conjunto de datos de morbilidad por cáncer entre los años 2019 y 2020 en el municipio de Pereira-Risaralda. Esto con la finalidad de pronosticar y caracterizar el tipo de población mas susceptible de padecer esta enfermedad según su edad, genero, zona y régimen de salud.

**Conjunto de datos.** Un total de 817 muestras de tumores de mama se perfilaron con cinco plataformas diferentes como se ha descrito previamente (Cancer Genome Atlas Research Network, 2014) y también se perfilaron 633 casos mediante matriz de proteínas en fase inversa (RPPA)


## Análisis exploratorio de datos

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

### BIRCH Clustering

with open('birch Cluster 0.csv') as f:
    Cluster_0_birch=pd.read_csv(f, delimiter=',')

with open('birch Cluster 1.csv') as f:
    Cluster_1_birch=pd.read_csv(f, delimiter=',')

with open('birch Cluster 2.csv') as f:
    Cluster_2_birch=pd.read_csv(f, delimiter=',')

with open('birch Cluster 3.csv') as f:
    Cluster_3_birch=pd.read_csv(f, delimiter=',')
	

#Cluster_0_birch = create_report(Cluster_0_birch, title='Cluster_0_birch')
#Cluster_1_birch = create_report(Cluster_1_birch, title='Cluster_1_birch')
#Cluster_2_birch = create_report(Cluster_2_birch, title='Cluster_2_birch')
#Cluster_3_birch = create_report(Cluster_3_birch, title='Cluster_3_birch')

#Cluster_0_birch.show_browser()
#Cluster_1_birch.show_browser()
#Cluster_2_birch.show_browser()
#Cluster_3_birch.show_browser()

### Compare Carcinoma BIRCH

import matplotlib.pyplot as plt
PATH_CLUSTERS = 'C:\\Users\\jamil\\OneDrive\\Escritorio\\DSM-BCD\\DSM-BCD\\NOTEBOOK\\IMAGENES_BIRCH_CLUSTERING\\'

variables = ['oncotree_code']
a=0
for i in variables:
        a= a +1
        categorical=sns.countplot(x=Cluster_0_birch[i], order = Cluster_0_birch[i].value_counts().iloc[:10].index, color='#8a93f8')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 0')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_0_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()	

        categorical=sns.countplot(x=Cluster_1_birch[i], order = Cluster_1_birch[i].value_counts().iloc[:10].index, color='#ec8273')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 1')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_1_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()

        categorical=sns.countplot(x=Cluster_2_birch[i], order = Cluster_2_birch[i].value_counts().iloc[:10].index, color='#00cc96')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 2')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_2_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()
        
        categorical=sns.countplot(x=Cluster_3_birch[i], order = Cluster_3_birch[i].value_counts().iloc[:10].index, color='#ab63fa')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 3')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_3_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()			


### Compare BIRCH

import matplotlib.pyplot as plt
PATH_CLUSTERS = 'C:\\Users\\jamil\\OneDrive\\Escritorio\\DSM-BCD\\DSM-BCD\\NOTEBOOK\\IMAGENES_BIRCH_CLUSTERING\\'

variables = ['oncotree_code','neoplasm_lymph_code','neoplasm_disease_stage_code','tumor_stage_code','her_2_ihc_percent_positive',
             'lymph_presentation','positive_lymph_hematoxylin','mutation_count','overall_survival_months',
             'tmb_nonsynonymous']

a=0
for i in variables:
    if Cluster_0_birch[i].dtypes == 'float64' or Cluster_0_birch[i].dtypes == 'int64':
        a= a +1
        numerical=sns.histplot(Cluster_0_birch[i],kde = True, color='#8a93f8')
        numerical.tick_params(axis='x', rotation=75)
        numerical.set(title='Cluster 0')
        plot_num=numerical.get_figure()
        plot_num.savefig(PATH_CLUSTERS+str(a)+'_Cluster_0_'+i+".pdf",bbox_inches = 'tight')
        plt.clf()

    if Cluster_0_birch[i].dtypes == 'object' or Cluster_0_birch[i].dtype== 'bool' : 
        a= a +1
        categorical=sns.countplot(x=Cluster_0_birch[i], order = Cluster_0_birch[i].value_counts().iloc[:10].index, color='#8a93f8')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 0')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_0_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()	


a=0
for i in variables:
    if Cluster_3_birch[i].dtypes == 'float64' or Cluster_3_birch[i].dtypes == 'int64':
        a= a +1
        numerical=sns.histplot(Cluster_3_birch[i],kde = True,color='#ab63fa')
        numerical.tick_params(axis='x', rotation=75)
        numerical.set(title='Cluster 3')
        plot_num=numerical.get_figure()
        plot_num.savefig(PATH_CLUSTERS+str(a)+'_Cluster_3_'+i+".pdf",bbox_inches = 'tight')
        plt.clf()

    if Cluster_3_birch[i].dtypes == 'object' or Cluster_3_birch[i].dtype== 'bool' : 
        a= a +1
        categorical=sns.countplot(x=Cluster_3_birch[i], order = Cluster_3_birch[i].value_counts().iloc[:10].index,color='#ab63fa')
        categorical.tick_params(axis='x', rotation=75)
        categorical.set(title='Cluster 3')
        plot_cat = categorical.get_figure()
        plot_cat.savefig(PATH_CLUSTERS+str(a)+'_Cluster_3_'+i+".pdf",bbox_inches = 'tight')    
        plt.clf()	

'''Cluster_0=Cluster_0_birch.applymap(str)
Cluster_1=Cluster_1_birch.applymap(str)
Cluster_2=Cluster_2_birch.applymap(str)
Cluster_3=Cluster_3_birch.applymap(str)

results = plot_diff([Cluster_0,Cluster_1,Cluster_2,Cluster_3],
config={
    "diff.label": ["Cluster 0", "Cluster 1" , "Cluster 2","Cluster 3"],
    "diff.baseline": 1,
    "diff.density": True
    }
)
results.show_browser()'''
