==> DSM-BCD/NOTEBOOK/ALGORITMOS/6_EDA_CLASSIFICATIONS_RESULTS.ipynb <==
# Data Science Methodology for Breast Cancer Diagnosis (DSM-BCD)

**Problema.** Colombia presenta limitaciones con respecto al acceso de la detección y el diagnóstico temprano del cáncer, provocado en la mayoría de los casos por factores como el estrato socio-económico, la cobertura del seguro de salud, el origen y la accesibilidad. En promedio, el tiempo de espera de un paciente es de 90 días desde la aparición de los síntomas hasta el diagnóstico de dicho cáncer. La primera acción para reducir la tasa de mortalidad por cáncer de mama debe estar enfocada en la agilidad del diagnóstico y el acceso oportuno a la atención.  Así, el objetivo de esta investigación es aplicar las etapas de la metodología KDD\footnote{Knowledge Discovery in Databases} al conjunto de datos de morbilidad por cáncer entre los años 2019 y 2020 en el municipio de Pereira-Risaralda. Esto con la finalidad de pronosticar y caracterizar el tipo de población mas susceptible de padecer esta enfermedad según su edad, genero, zona y régimen de salud.

**Conjunto de datos.** Un total de 817 muestras de tumores de mama se perfilaron con cinco plataformas diferentes como se ha descrito previamente (Cancer Genome Atlas Research Network, 2014) y también se perfilaron 633 casos mediante matriz de proteínas en fase inversa (RPPA)


## Análisis exploratorio de datos

import pandas as pd
import numpy as np
from IPython.display import Image
import matplotlib.pyplot as plt
from dataprep.eda import plot, plot_correlation, plot_missing,create_report
from dataprep.clean import clean_headers
from dataprep.clean import clean_duplication
from dataprep.eda import plot_diff

with open('BRCA.csv') as f:
    BRCA=pd.read_csv(f, delimiter=',')

with open('IDC.csv') as f:
    IDC=pd.read_csv(f, delimiter=',')

with open('ILC.csv') as f:
    ILC=pd.read_csv(f, delimiter=',')

with open('MDLC.csv') as f:
    MDLC=pd.read_csv(f, delimiter=',')

BRCA = create_report(BRCA, title='BRCA')
BRCA.show_browser()

IDC = create_report(IDC, title='IDC')
IDC.show_browser()

ILC = create_report(ILC, title='ILC')
ILC.show_browser()

MDLC = create_report(MDLC, title='MDLC')
MDLC.show_browser()


BRCA=BRCA.applymap(str)
IDC=IDC.applymap(str)
ILC=ILC.applymap(str)
MDLC=MDLC.applymap(str)

results = plot_diff([BRCA,IDC,ILC,MDLC],
config={
    "diff.label": ["BRCA", "IDC" , "ILC", "MDLC" ],
    "diff.baseline": 1,
    "diff.density": True
    }
)
#results.show_browser()
